fix: cast thinker_hidden_states to float32 for serialization (#181)
vllm-omni's orchestrator serializes additional_information via numpy,
which doesn't support bfloat16. Cast to float32 before passing between
thinker→talker stages.

Also tune stage_configs for single-GPU RTX 3090 deployment
(0.80+0.07+0.02 = 0.89 gpu_memory_utilization).

+++ vllm_omni/model_executor/stage_input_processors/minicpm_o.py
@@ -132,7 +132,8 @@ def thinker2talker(
             "thinker_token_ids": torch.tensor(
                 tts_token_ids, dtype=torch.long, device=device,
             ),
-            "thinker_hidden_states": tts_hidden.detach().to(device=device),
+            # Cast to float32 for serialization (numpy doesn't support bfloat16)
+            "thinker_hidden_states": tts_hidden.detach().float().to(device=device),
         }
 
         # +2 for text_eos and audio_bos boundary tokens appended in talker_preprocess
